feat(Height): 원시 값 포장
